command invoker does not support streams

+++ .github/workflows/run.sh
@@ -23,9 +23,9 @@ fi
 
 # workaround for https://github.com/projectriff/node-function-invoker/issues/113
 if [ ${CLUSTER} = "pks-gcp" ]; then
-  languages="command java java-boot"
+  languages="java java-boot command"
 else
-  languages="command node npm java java-boot"
+  languages="node npm java java-boot command"
 fi
 for mode in ${modes}; do
   # functions
@@ -57,52 +57,54 @@ for mode in ${modes}; do
     riff knative deployer delete ${name} --namespace ${NAMESPACE}
 
     # streaming runtime
-    lower_stream=${name}-lower
-    upper_stream=${name}-upper
+    if [ "$test" != "commnd" ]; then
+      lower_stream=${name}-lower
+      upper_stream=${name}-upper
 
-    riff streaming stream create ${lower_stream} --namespace $NAMESPACE --provider franz-kafka-provisioner --content-type 'text/plain'
-    riff streaming stream create ${upper_stream} --namespace $NAMESPACE --provider franz-kafka-provisioner --content-type 'text/plain'
+      riff streaming stream create ${lower_stream} --namespace $NAMESPACE --provider franz-kafka-provisioner --content-type 'text/plain'
+      riff streaming stream create ${upper_stream} --namespace $NAMESPACE --provider franz-kafka-provisioner --content-type 'text/plain'
 
-    # TODO remove once riff streaming stream supports --tail
-    kubectl wait streams.streaming.projectriff.io ${lower_stream} --for=condition=Ready --namespace $NAMESPACE --timeout=60s
-    kubectl wait streams.streaming.projectriff.io ${upper_stream} --for=condition=Ready --namespace $NAMESPACE --timeout=60s
+      # TODO remove once riff streaming stream supports --tail
+      kubectl wait streams.streaming.projectriff.io ${lower_stream} --for=condition=Ready --namespace $NAMESPACE --timeout=60s
+      kubectl wait streams.streaming.projectriff.io ${upper_stream} --for=condition=Ready --namespace $NAMESPACE --timeout=60s
 
-    riff streaming processor create $name --function-ref $name --namespace $NAMESPACE --input ${lower_stream} --output ${upper_stream}
-    riff streaming processor tail $name --namespace $NAMESPACE > processor.${name}.log &
-    processor_pid=$?
-    kubectl wait processors.streaming.projectriff.io $name --for=condition=Ready --namespace $NAMESPACE --timeout=60s
+      riff streaming processor create $name --function-ref $name --namespace $NAMESPACE --input ${lower_stream} --output ${upper_stream}
+      riff streaming processor tail $name --namespace $NAMESPACE > processor.${name}.log &
+      processor_pid=$?
+      kubectl wait processors.streaming.projectriff.io $name --for=condition=Ready --namespace $NAMESPACE --timeout=60s
 
-    kubectl exec riff-dev -n $NAMESPACE -- subscribe ${upper_stream} -n $NAMESPACE --payload-as-string | tee result.txt &
-    kubectl exec riff-dev -n $NAMESPACE -- publish ${lower_stream} -n $NAMESPACE --payload "fats" --content-type "text/plain"
+      kubectl exec riff-dev -n $NAMESPACE -- subscribe ${upper_stream} -n $NAMESPACE --payload-as-string | tee result.txt &
+      kubectl exec riff-dev -n $NAMESPACE -- publish ${lower_stream} -n $NAMESPACE --payload "fats" --content-type "text/plain"
 
-    actual_data=""
-    expected_data="FATS"
-    cnt=1
-    while [ $cnt -lt 60 ]; do
-      echo -n "."
-      cnt=$((cnt+1))
+      actual_data=""
+      expected_data="FATS"
+      cnt=1
+      while [ $cnt -lt 60 ]; do
+        echo -n "."
+        cnt=$((cnt+1))
 
-      actual_data=`cat result.txt | jq -r .payload`
-      if [ "$actual_data" == "$expected_data" ]; then
-        break
-      fi
+        actual_data=`cat result.txt | jq -r .payload`
+        if [ "$actual_data" == "$expected_data" ]; then
+          break
+        fi
 
-      sleep 1
-    done
+        sleep 1
+      done
 
-    echo ""
-    echo "Processor log:"
-    cat processor.${name}.log
-    echo ""
-    # kill $processor_pid
+      echo ""
+      echo "Processor log:"
+      cat processor.${name}.log
+      echo ""
+      # kill $processor_pid
 
-    fats_assert "$expected_data" "$actual_data"
+      fats_assert "$expected_data" "$actual_data"
 
-    kubectl exec riff-dev -n $NAMESPACE -- sh -c 'kill $(pidof subscribe)'
+      kubectl exec riff-dev -n $NAMESPACE -- sh -c 'kill $(pidof subscribe)'
 
-    riff streaming stream delete ${lower_stream} --namespace $NAMESPACE
-    riff streaming stream delete ${upper_stream} --namespace $NAMESPACE
-    riff streaming processor delete $name --namespace $NAMESPACE
+      riff streaming stream delete ${lower_stream} --namespace $NAMESPACE
+      riff streaming stream delete ${upper_stream} --namespace $NAMESPACE
+      riff streaming processor delete $name --namespace $NAMESPACE
+    fi
 
     # cleanup
     riff function delete ${name} --namespace ${NAMESPACE}
